\name{hidetify}
\alias{hidetify}
\alias{hidetify}
\docType{package}
\title{
\packageTitle{hidetify}
}
\description{
\packageDescription{hidetify}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hidetify}
\packageIndices{hidetify}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{    
\packageAuthor{hidetify}

Maintainer: \packageMaintainer{hidetify}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
