% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hidetify_data.R
\docType{data}
\name{sim_hidetify_data}
\alias{sim_hidetify_data}
\title{sim_hidetify_data for the hidetify package}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{y}{The original response variable}
 \item{youtlier}{The contaminated response variable, where the first 10 observations are contaminated}
 \item{x1}{The first predictor}
 \item{x1000}{The last predictor. There is 1000 predictors in total.}
}
}
\usage{
data(sim_hidetify_data)
}
\description{
Simulated toy data that comes with the hidetify package.
The data is generated according to model I of the single detection paper
where only the response variable is contaminated.
The data set has 1002 variables. The first variable y is the original response variable.
The second variable youtlier is the contaminated response variable, where the first 10 observations 
are contaminated. The rest of the variables (1000) are the predictors.
}
\examples{

data(sim_hidetify_data)
head(sim_hidetify_data)

}
\references{
This data set was artificially created for the hidetify package.
}
\keyword{datasets}
